requirements.txt
# DriftBench-TS Requirements

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
PyYAML>=6.0
scipy>=1.10.0
matplotlib>=3.7.0
tqdm>=4.65.0

# ML models
lightgbm>=4.0.0
torch>=2.0.0

# Dashboard
flask>=3.0.0
